refactor/transaction-balance (#20)

+++ src/db/models/member.ts
@@ -67,8 +67,8 @@ Member.init(
       type: DataTypes.STRING
     },
     balance: {
-      type: DataTypes.DECIMAL(10, 2),
-      defaultValue: 0.0
+      type: DataTypes.INTEGER,
+      defaultValue: 0
     },
     status: {
       type: DataTypes.STRING,
